Moved plot_Nd to shared location, added function documentation, added library requirements
