+++ seismic/inversion/wavefield_decomp/requirements.txt
@@ -1,1 +1,3 @@
 sortedcontainers
+seaborn
+landscapes
